# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ChatMessageRole = typeof ChatMessageRole[keyof typeof ChatMessageRole];


export const ChatMessageRole = {
  system: 'system',
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatInput {
  messages: ChatMessage[];
  /**
     * @minimum 0
     * @maximum 1.2
     */
  temperature?: number;
  /**
     * @minimum 256
     * @maximum 4000
     */
  maxTokens?: number;
}

export interface ChatResponse {
  ok: boolean;
  /** @nullable */
  text: string | null;
  /** @nullable */
  thought: string | null;
  /** @nullable */
  error: string | null;
}

export type CurrentUserWorkspaceType = typeof CurrentUserWorkspaceType[keyof typeof CurrentUserWorkspaceType];


export const CurrentUserWorkspaceType = {
  'managed-sandbox': 'managed-sandbox',
} as const;

export interface CurrentUser {
  id: string;
  displayName: string;
  /** @nullable */
  email: string | null;
  buildCreditsRemaining: number;
  workspaceId: string;
  workspaceType: CurrentUserWorkspaceType;
}

export type ThreadMode = typeof ThreadMode[keyof typeof ThreadMode];


export const ThreadMode = {
  chat: 'chat',
  build: 'build',
} as const;

export type ThreadMessageRole = typeof ThreadMessageRole[keyof typeof ThreadMessageRole];


export const ThreadMessageRole = {
  system: 'system',
  user: 'user',
  assistant: 'assistant',
} as const;

export interface ThreadMessage {
  id: string;
  role: ThreadMessageRole;
  content: string;
  /** @nullable */
  thought: string | null;
  createdAt: string;
}

export interface Thread {
  id: string;
  mode: ThreadMode;
  title: string;
  createdAt: string;
  updatedAt: string;
  messages: ThreadMessage[];
}

export type ThreadInputMode = typeof ThreadInputMode[keyof typeof ThreadInputMode];


export const ThreadInputMode = {
  chat: 'chat',
  build: 'build',
} as const;

export interface ThreadInput {
  mode: ThreadInputMode;
  title: string;
  createdAt: string;
  messages: ThreadMessage[];
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  building: 'building',
  ready: 'ready',
  failed: 'failed',
} as const;

export type ProjectServerStatus = typeof ProjectServerStatus[keyof typeof ProjectServerStatus];


export const ProjectServerStatus = {
  stopped: 'stopped',
  starting: 'starting',
  running: 'running',
  failed: 'failed',
} as const;

export interface Project {
  id: string;
  name: string;
  prompt: string;
  status: ProjectStatus;
  serverStatus: ProjectServerStatus;
  /** @nullable */
  serverPort: number | null;
  /** @nullable */
  buildLog: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface BuildInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  prompt: string;
}

export interface EditProjectInput {
  /** @minLength 1 */
  instruction: string;
}

export type ProjectFileEntryKind = typeof ProjectFileEntryKind[keyof typeof ProjectFileEntryKind];


export const ProjectFileEntryKind = {
  file: 'file',
  directory: 'directory',
} as const;

export interface ProjectFileEntry {
  path: string;
  kind: ProjectFileEntryKind;
  size: number;
}

export interface ProjectFile {
  path: string;
  content: string;
  size: number;
}

export interface ProjectFileInput {
  /** @minLength 1 */
  path: string;
}

export type ServerActionInputAction = typeof ServerActionInputAction[keyof typeof ServerActionInputAction];


export const ServerActionInputAction = {
  start: 'start',
  stop: 'stop',
} as const;

export interface ServerActionInput {
  action: ServerActionInputAction;
}