# lib/api-zod/src/generated/types/buildInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BuildInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  prompt: string;
}