# lib/api-zod/src/generated/types/projectFileEntry.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProjectFileEntryKind } from './projectFileEntryKind';

export interface ProjectFileEntry {
  path: string;
  kind: ProjectFileEntryKind;
  size: number;
}