# lib/api-zod/src/generated/types/serverActionInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ServerActionInputAction } from './serverActionInputAction';

export interface ServerActionInput {
  action: ServerActionInputAction;
}