# lib/api-zod/src/generated/types/threadMessage.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ThreadMessageRole } from './threadMessageRole';

export interface ThreadMessage {
  id: string;
  role: ThreadMessageRole;
  content: string;
  /** @nullable */
  thought: string | null;
  createdAt: Date;
}